pls fix 10

+++ tools/In_Cluster_GHAR/runner_setup.sh
@@ -23,17 +23,11 @@ function deploy_runner()
 
 }
 
-function restart_secret()
-{
-	if kubectl get deployment github-runner -n actions-runner; then
-		if kubectl delete secret github-runner-secrets --namespace actions-runner; then
+if kubectl get deployment github-runner -n actions-runner && kubectl get secret github-runner-secrets -n actions-runner; then
+	if kubectl delete secret github-runner-secrets --namespace actions-runner; then
 			kubectl create secret generic github-runner-secrets --from-literal=GITHUB_TOKEN=$token --from-literal=GITHUB_URL="https://github.com/$org/$repo" -n actions-runner
-		fi
 	fi
-}
-
-restart_secret()
-if kubectl delete deployment github-runner --namespace=actions-runner; then
+elif kubectl delete deployment github-runner --namespace=actions-runner; then
 	if kubectl delete secret github-runner-secrets --namespace actions-runner; then
 			echo "Reinstalling the github-runner"
 			deploy_runner
@@ -51,6 +45,5 @@ else
 		echo "Creating new in cluster runner"
 		deploy_runner
 	fi
-fi
 
-
+fi
